chore(cli): Reorganize modules
