+++ src/commands/completions/zsh.rs
@@ -137,7 +137,7 @@ fn escape(s: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::completions::fake_pr_command;
+    use crate::commands::completions::fake_pr_command;
 
     #[test]
     fn compdef_and_chains() {
