+++ src/lib.rs
@@ -1,17 +1,16 @@
-use clap::CommandFactory as _;
-
 mod auth;
 mod cli;
-mod completions;
+mod commands;
 pub mod config;
-mod debug;
+mod editor;
+mod frontmatter;
 mod fs;
 mod gh;
 mod git;
 mod jj;
 mod macro_support;
-mod pr;
 mod proc;
+mod template;
 mod ui;
 mod util;
 
@@ -31,25 +30,13 @@ pub async fn dispatch(bin_name: &str) -> anyhow::Result<()> {
     let _logger = logging::init(args.global.resolve_log_level())?;
     let global = args.global;
     match args.command {
-        Command::Pr { action } => pr::dispatch(global, action).await?,
-        Command::Debug { action } => debug::dispatch(global, action).await?,
+        Command::Pr { action } => commands::pr::dispatch(global, action).await?,
+        Command::Debug { action } => commands::debug::dispatch(global, action).await?,
         Command::Completions {
             shell,
             jj_alias,
             jj_gh_subcommand,
-        } => match (jj_alias, jj_gh_subcommand) {
-            (Some(alias), Some(subcommand)) => {
-                completions::run(shell.into(), &alias, subcommand, &mut std::io::stdout())?;
-            }
-            _ => {
-                clap_complete::generate(
-                    shell,
-                    &mut Cli::command(),
-                    bin_name,
-                    &mut std::io::stdout(),
-                );
-            }
-        },
+        } => commands::completions::run(bin_name, shell, jj_alias, jj_gh_subcommand)?,
     }
     Ok(())
 }
